Fix type assertion for CTA overlay request in beforevalidationhook.go to ensure correct handling of float64 values.

+++ modules/pubmatic/openwrap/beforevalidationhook.go
@@ -648,7 +648,7 @@ func (m OpenWrap) handleBeforeValidationHook(
 			}
 		}
 
-		isCTAOverlayRequest := impExt.OWSDK != nil && impExt.OWSDK["ctaoverlay"] == 1
+		isCTAOverlayRequest := impExt.OWSDK != nil && impExt.OWSDK["ctaoverlay"] == float64(1)
 		impExt.Wrapper = nil
 		impExt.Reward = nil
 		impExt.Bidder = nil
